+++ gatsby-node.js
@@ -9,6 +9,27 @@ const kebabCase = (string) =>
     .replace(/\s+/g, '-')
     .toLowerCase();
 
+// This is a temporary customization to ensure that the path field is always
+// queryable from the createPages API. We rely on the path for deprecation
+// warnings to inform authors that path is no longer a supported markdown
+// property. Because of Gatsby's type inference from frontmatter properties, the
+// query in createPages fails without this patch if no file contains a 'path'
+// frontmatter property.
+//
+// This patch can be safely removed when removing the deprecation warning in
+// createPages.
+exports.createSchemaCustomization = ({ actions }) => {
+  const { createTypes } = actions;
+
+  const typeDefs = `
+    type MdxFrontmatter {
+      path: String
+    }
+  `;
+
+  createTypes(typeDefs);
+};
+
 exports.createPages = async ({ actions, graphql, reporter }) => {
   const { createPage, createRedirect } = actions;
 
@@ -89,25 +110,46 @@ exports.createPages = async ({ actions, graphql, reporter }) => {
       frontmatter.redirects.forEach((fromPath) => {
         createRedirect({
           fromPath,
-          toPath: frontmatter.path,
+          toPath: slug,
           isPermanent: true,
           redirectInBrowser: true,
         });
       });
     }
 
+    if (frontmatter.path) {
+      const recommendedPath = path.join(
+        'src/markdown-pages',
+        fileRelativePath.endsWith('index.mdx')
+          ? path.join(frontmatter.path, 'index.mdx')
+          : `${frontmatter.path}.mdx`
+      );
+
+      const recommendation =
+        recommendedPath === fileRelativePath
+          ? "Please remove the 'path' property."
+          : `To render this page at '${frontmatter.path}', please move this file to: '${recommendedPath}'.`;
+
+      reporter.panicOnBuild(
+        `
+${fileRelativePath}
+
+The 'path' property on frontmatter is deprecated and has no effect. URLs are now generated using the path of the file. ${recommendation}
+`.trim()
+      );
+    }
+
     createPage({
-      path: frontmatter.path
-        ? path.join(frontmatter.path, '/')
-        : path.join(slug, '/'),
+      path: path.join(slug, '/'),
       component: path.resolve(`src/templates/${frontmatter.template}.js`),
       context: {
         slug,
         fileRelativePath,
-        guidesFilter:
-          ['OverviewTemplate', 'LabOverviewTemplate'].includes(frontmatter.template)
-            ? `${frontmatter.path}/*`
-            : undefined,
+        guidesFilter: ['OverviewTemplate', 'LabOverviewTemplate'].includes(
+          frontmatter.template
+        )
+          ? `${slug}/*`
+          : undefined,
       },
     });
   });