+++ src/pages/instant-observability.js
@@ -22,8 +22,6 @@ import BUILD_YOUR_OWN from '../images/build-your-own.svg';
 import GUIDED_INSTALL from '../images/guided-install.svg';
 import { useDebounce } from 'react-use';
 import { sortFeaturedQuickstarts } from '../utils/sortFeaturedQuickstarts';
-import SuperTilesExperiment from '../experiments/super_tiles';
-
 import {
   QUICKSTARTS_REPO,
   RESERVED_QUICKSTART_IDS,
@@ -32,6 +30,7 @@ import {
 import CATEGORIES from '../data/instant-observability-categories';
 
 import { getGuidedInstallStackedNr1Url } from '../utils/get-pack-nr1-url';
+import SuperTiles from '../components/SuperTiles';
 
 const FILTERS = [
   { name: 'Dashboards', type: 'dashboards', icon: 'nr-dashboard' },
@@ -296,7 +295,6 @@ const QuickstartsPage = ({ data, location }) => {
             }
           `}
         >
-          {isMobile && <SuperTilesExperiment isMobile />}
           <div
             css={css`
               padding: var(--site-content-padding);
@@ -412,7 +410,7 @@ const QuickstartsPage = ({ data, location }) => {
             padding: var(--site-content-padding);
           `}
         >
-          {!isMobile && <SuperTilesExperiment />}
+          <SuperTiles />
           <div
             css={css`
               background-color: var(--secondary-background-color);